refactor(UnstyledTextInput): moved to internal (#3376)
